Provide error default for redshift secgroups
Instead of skipping the redshift test by default,
if there is an error (most likely due to using EC2-VPC)
the resources call returns an empty list. Pytest will
display this as a `skip`, which is what we would want.

The main thing this provides is that it will support both
EC2-VPC and EC2-Classic without any code/config changes.

Fixes #66