Fix the concept of "admin" in aws iam tests (#152)
Originally, a user was an "admin" if they had a policy that included the
word "admin" in it. That kinda worked sometimes, but needed to be
replaced.

Now, in the config file you can list the admin policies and groups and a
user or role will be matched against these lists.