Search for metadata in funcargs
Code previously only searched for metadata within test arguments that
were dictionaries.

For GitHub, the dictionary-of-arguements (aka funcargs) is where the
meta data lives. To convert scalar args to dictionaries would also
obfuscate paramertized name of the function call (a nice pytest feature,
imo).

Searching funcargs is done in addition to, and previous to, the search
into dictionaries. That way, if dictionaries are present, they will
override the scalar args.