Makefile should report non prod python versions (#305)
Fixes #304

We really care about the minimum version that is currently advertised as
being supported. The assumption is, anything that works on the minimum
version will continue to work on a later version.

Yes, CI will catch this -- the hope is this message may prevent a dev
from only finding out that late in the process. We could consider
removing this once tox testing is implemented.