Rename, as we need to invoke during production run
We want to run the data collection steps separately, to provide the raw
data for dashboards.

The eventual plan is to have the github data gathering be a separate
package. Until then, we fake it by having a wrapper script to align all
the imports. The script also helps when trying to debug, especially from
vs-code.